+++ crates/ika-core/src/dwallet_mpc/integration_tests/create_dwallet.rs
@@ -104,6 +104,14 @@ async fn dwallet_dkg_first_round() {
     info!("DWallet DKG first round completed");
 }
 
+pub(crate) struct DWalletTestResult {
+    pub(crate) flow_completion_consensus_round: Round,
+    pub(crate) dkg_second_round_output: DKGSecondRoundOutput,
+    pub(crate) dwallet_secret_key_share: Vec<u8>,
+    pub(crate) class_groups_encryption_key: Vec<u8>,
+    pub(crate) class_groups_decryption_key: Vec<u8>,
+}
+
 #[tokio::test]
 #[cfg(test)]
 /// Runs a network DKG and then uses the resulting network key to run the DWallet DKG first round.
@@ -140,13 +148,44 @@ async fn create_dwallet() {
         create_dwallet_test(&mut test_state, consensus_round, key_id, network_key_bytes).await;
     info!("DWallet DKG second round completed");
 }
-pub(crate) struct DWalletTestResult {
-    pub(crate) flow_completion_consensus_round: Round,
-    pub(crate) dkg_second_round_output: DKGSecondRoundOutput,
-    pub(crate) dwallet_secret_key_share: Vec<u8>,
-    pub(crate) class_groups_encryption_key: Vec<u8>,
-    pub(crate) class_groups_decryption_key: Vec<u8>,
+
+#[tokio::test]
+#[cfg(test)]
+/// Runs a network DKG and then uses the resulting network key to run the DWallet DKG first round.
+async fn create_imported_dwallet() {
+    let _ = tracing_subscriber::fmt().with_test_writer().try_init();
+    let (committee, _) = Committee::new_simple_test_committee();
+    let ika_network_config = IkaNetworkConfig::new_for_testing();
+    let epoch_id = 1;
+    let (
+        mut dwallet_mpc_services,
+        mut sui_data_senders,
+        mut sent_consensus_messages_collectors,
+        mut epoch_stores,
+        notify_services,
+    ) = utils::create_dwallet_mpc_services(4);
+    let mut test_state = IntegrationTestState {
+        dwallet_mpc_services,
+        sent_consensus_messages_collectors,
+        epoch_stores,
+        notify_services,
+        crypto_round: 1,
+        consensus_round: 1,
+        committee,
+        sui_data_senders,
+    };
+    for service in &mut test_state.dwallet_mpc_services {
+        service
+            .dwallet_mpc_manager_mut()
+            .last_session_to_complete_in_current_epoch = 4;
+    }
+    let (consensus_round, network_key_bytes, key_id) =
+        create_network_key_test(&mut test_state).await;
+    let result =
+        create_dwallet_test(&mut test_state, consensus_round, key_id, network_key_bytes).await;
+    info!("DWallet DKG second round completed");
 }
+
 pub(crate) async fn create_dwallet_test(
     mut test_state: &mut IntegrationTestState,
     start_consensus_round: Round,
