+++ crates/ika-core/src/dwallet_mpc/integration_tests/malicious_behavior.rs
@@ -8,6 +8,8 @@ use ika_types::messages_dwallet_mpc::{
 use itertools::Itertools;
 use std::collections::HashMap;
 use tracing::info;
+use std::sync::Arc;
+use crate::dwallet_mpc::integration_tests::utils::TestingSubmitToConsensus;
 
 #[tokio::test]
 #[cfg(test)]
@@ -176,7 +178,7 @@ async fn test_party_copies_other_party_message_dkg_round() {
     .await;
 
     for (copying_party, copied_party) in copying_parties.iter() {
-        utils::replace_party_message_with_other_party_message(
+        replace_party_message_with_other_party_message(
             *copying_party as usize,
             *copied_party as usize,
             mpc_round,
@@ -226,4 +228,41 @@ async fn test_party_copies_other_party_message_dkg_round() {
             malicious_actor_name
         );
     }
+}
+
+pub(crate) fn replace_party_message_with_other_party_message(
+    party_to_replace: usize,
+    other_party: usize,
+    crypto_round: u64,
+    sent_consensus_messages_collectors: &mut Vec<Arc<TestingSubmitToConsensus>>,
+) {
+    let original_message = sent_consensus_messages_collectors[party_to_replace]
+        .submitted_messages
+        .lock()
+        .unwrap()
+        .pop()
+        .unwrap();
+
+    let mut other_party_message = sent_consensus_messages_collectors[other_party]
+        .submitted_messages
+        .lock()
+        .unwrap()
+        .first()
+        .unwrap()
+        .clone();
+    let ConsensusTransactionKind::DWalletMPCMessage(ref mut other_party_message_content) =
+        other_party_message.kind
+    else {
+        panic!("Only DWalletMPCMessage messages can be replaced with other party messages");
+    };
+    let ConsensusTransactionKind::DWalletMPCMessage(mut original_message) = original_message.kind
+    else {
+        panic!("Only DWalletMPCMessage messages can be replaced with other party messages");
+    };
+    other_party_message_content.authority = original_message.authority;
+    sent_consensus_messages_collectors[party_to_replace]
+        .submitted_messages
+        .lock()
+        .unwrap()
+        .push(other_party_message)
 }
